abstract default chain generation into util for testing
