+++ config_yourself/util.py
@@ -18,6 +18,8 @@ except ImportError:  # pragma: no cover
         # noinspection PyPackageRequirements
         from future.moves.collections import Mapping, Hashable
 import warnings
+from os import environ, path
+from config_yourself.exceptions import InvalidConfig
 
 
 def merge_dicts(src, dest, parent=[]):
@@ -45,3 +47,59 @@ def merge_dicts(src, dest, parent=[]):
             warnings.warn(msg, DeprecationWarning)
             pass
     return src
+
+
+def _default_config_chain(
+    config_file=None, before=None, after=None, config_folder="./config"
+):
+    """Generate a chain of config file path names
+
+    With no arguments, this means:
+
+    * the ``"{config_folder}/defaults.yml"`` file,
+    * a file located at ``os.environ.get("CONFIG_FILE", path.join(config_folder, "local.yml"))``
+    * a ``"{config_folder}/personal.yml"`` file, if it exists
+
+    :param Optional[str] config_file: the path to the main config file to load, usually in the form of ``./config/your-environment-name.yml``. (default: ``os.environ.get("CONFIG_FILE", "./config/local.yml")``)
+    :param Optional[list[str]] before: File names to load before ``config_file`` without (default: `['default']`)
+    :param Optional[list[str]] after: File names to apply after ``config_file`` if they exist (default: `['personal']`)
+    :param Optional[str] config_folder: Base folder path to look for files in, (default: `./config`)
+
+    :returns: A list of path names
+    :rtype: list[str]
+    """
+
+    if config_file is None:
+        config_file = environ.get("CONFIG_FILE", path.join(config_folder, "local.yml"))
+        if not path.isfile(config_file):
+            # Help users debug missing/mistyped CONFIG_FILEs
+            raise InvalidConfig("CONFIG_FILE='{}' is not a file".format(config_file))
+
+    if before is None:
+        before = ["defaults"]
+
+    if after is None:
+        after = ["personal"]
+
+    # the chain of files to merge before decrypting
+    chain = []
+    # extract some details about the main file
+    config_file = path.abspath(config_file)
+    if config_folder is None:
+        config_folder = path.dirname(config_file)
+    _main_ext = path.splitext(config_file)[1]
+
+    def name_to_path(name):
+        if not name.endswith(".yml") and not name.endswith(".yaml"):
+            name = "{}{}".format(name, _main_ext)
+        path.join(config_folder, name)
+
+    # Add the resolved before files
+    chain = [name_to_path(file_name) for file_name in before]
+    # Add the main config file
+    chain.append(config_file)
+
+    # Add after files if they exist
+    for f in [name_to_path(file_name) for file_name in after]:
+        if path.exists(f):
+            chain.append(f)
